Moved header files from dedicated include dir back into src/ (nevermind this idea; CMake packs the ed25519.h in shared builds into the nice output include dir just fine on its own).
Modified CHANGELOG accordingly.

Copy license.txt into CMake build output dir on package.

Added "ORLP_" prefix to ed25519.h header guard macro to prevent potential name clashes.
