+++ src/key_convert.c
@@ -1,5 +1,5 @@
-#include "../include/orlp-ed25519/ed25519.h"
-#include "../include/orlp-ed25519/sha512.h"
+#include "ed25519.h"
+#include "sha512.h"
 
 void ed25519_key_convert_ref10_to_orlp(const unsigned char *private_key_ref10, unsigned char *private_key_orlp) {
   sha512(private_key_ref10, 32, private_key_orlp);
