+++ src/keypair.c
@@ -1,6 +1,6 @@
-#include "../include/orlp-ed25519/ed25519.h"
-#include "../include/orlp-ed25519/sha512.h"
-#include "../include/orlp-ed25519/ge.h"
+#include "ed25519.h"
+#include "sha512.h"
+#include "ge.h"
 
 
 void ed25519_create_keypair(unsigned char *public_key, unsigned char *private_key, const unsigned char *seed) {
