+++ src/sc.h
