+++ src/sign.c
@@ -1,7 +1,7 @@
-#include "../include/orlp-ed25519/ed25519.h"
-#include "../include/orlp-ed25519/sha512.h"
-#include "../include/orlp-ed25519/ge.h"
-#include "../include/orlp-ed25519/sc.h"
+#include "ed25519.h"
+#include "sha512.h"
+#include "ge.h"
+#include "sc.h"
 
 
 void ed25519_sign(unsigned char *signature, const unsigned char *message, size_t message_len, const unsigned char *public_key, const unsigned char *private_key) {
